refactor(async-kv): merge MooncakeAsyncKVManager into base manager
Implement layerwise async KV as an optional path in MooncakeKVManager and expose a standard scheduler hook, removing the async-manager subclass and ad-hoc capability checks.
